feat: fully support vite plugin vue
