+++ packages/core/src/config/types.ts
@@ -56,9 +56,9 @@ export interface UserConfig {
   resolveConfigPath?: string;
   configFileDependencies?: string[];
   /** js plugin(which is a javascript object) and rust plugin(which is string refer to a .farm file or a package) */
-  plugins?: (RustPlugin | JsPlugin)[];
+  plugins?: (RustPlugin | JsPlugin)[] | JsPlugin[][];
   /** vite plugins */
-  vitePlugins?: any[];
+  vitePlugins?: (object | (() => { vitePlugin: any; filters: string[] }))[];
   /** config related to compilation */
   compilation?: Config['config'];
   /** config related to dev server */
